Moved SEE to ChessMechanics
