+++ src/BestMoveSearch.cpp
@@ -615,90 +615,3 @@ int BestMoveSearch::QuiesceEval()
 
     return _qSearch<SearchType::PVSearch>(_board, NEGATIVE_INFINITY, POSITIVE_INFINITY, hash, 0);
 }
-
-BestMoveSearch::_seePackage BestMoveSearch::_prepareForSEE(const int msbPos) const {
-    const uint64_t bishops = _board.BitBoards[wQueensIndex] | _board.BitBoards[bQueensIndex] |
-            _board.BitBoards[wBishopsIndex] | _board.BitBoards[bBishopsIndex];
-    const uint64_t rooks =  _board.BitBoards[wQueensIndex] | _board.BitBoards[bQueensIndex] |
-                            _board.BitBoards[wRooksIndex] | _board.BitBoards[bRooksIndex];
-    const uint64_t knights = _board.BitBoards[wKnightsIndex] | _board.BitBoards[bKnightsIndex];
-    const uint64_t kings = _board.BitBoards[wKingIndex] | _board.BitBoards[bKnightsIndex];
-    const uint64_t fullMap = bishops | rooks | knights | kings | _board.BitBoards[wPawnsIndex] |
-            _board.BitBoards[bPawnsIndex];
-
-    uint64_t attackers = 0;
-    attackers |= (BishopMap::GetMoves(msbPos, fullMap) & bishops)
-                    | (RookMap::GetMoves(msbPos, fullMap) & rooks)
-                    | (KnightMap::GetMoves(msbPos) & knights)
-                    | (KingMap::GetMoves(msbPos) & kings);
-
-    const uint64_t figBitBoard = MaxMsbPossible >> msbPos;
-    attackers |= (WhitePawnMap::GetAttackFields(figBitBoard) & _board.BitBoards[bPawnsIndex])
-            | (BlackPawnMap::GetAttackFields(figBitBoard) & _board.BitBoards[wPawnsIndex]);
-
-    const uint64_t mayXray = bishops | rooks | _board.BitBoards[wPawnsIndex] | _board.BitBoards[bPawnsIndex];
-    return {attackers, fullMap, mayXray};
-}
-
-uint64_t BestMoveSearch::_updateAttackers(const uint64_t fullMap, const int msbPos) const {
-    const uint64_t bishops = (_board.BitBoards[wQueensIndex] | _board.BitBoards[bQueensIndex] |
-                             _board.BitBoards[wBishopsIndex] | _board.BitBoards[bBishopsIndex]) & fullMap;
-    const uint64_t rooks =  (_board.BitBoards[wQueensIndex] | _board.BitBoards[bQueensIndex] |
-                            _board.BitBoards[wRooksIndex] | _board.BitBoards[bRooksIndex]) & fullMap;
-
-    uint64_t attackers = 0;
-    attackers |= (BishopMap::GetMoves(msbPos, fullMap) & bishops)
-                 | (RookMap::GetMoves(msbPos, fullMap) & rooks);
-
-    return attackers;
-}
-
-int BestMoveSearch::_see(const Move mv) const {
-    static constexpr size_t MaximalFigureCount = 32;
-
-    int scores[MaximalFigureCount];
-    int depth = 0;
-    uint64_t attackFromBitBoard = MaxMsbPossible >> mv.GetStartField();
-    auto [attackersBitBoard, fullMap, xray] = _prepareForSEE(mv.GetTargetField());
-
-    int attackerFigType = FindFigType(attackFromBitBoard, _board);
-    int color = _board.MovingColor;
-    scores[depth] = BoardEvaluator::BasicFigureValues[mv.GetKilledBoardIndex()];
-    do {
-        depth++;
-
-        // sum up points
-        scores[depth] = BoardEvaluator::BasicFigureValues[attackerFigType] - scores[depth - 1];
-
-        // pseudo make move
-        attackersBitBoard ^= attackFromBitBoard;
-        fullMap ^= attackFromBitBoard;
-
-        // we moved a figure that could increase possible attacks counts on given field
-        if (attackFromBitBoard & xray)
-            attackersBitBoard |= _updateAttackers(fullMap, mv.GetTargetField());
-
-        color = SwapColor(color);
-        attackFromBitBoard = _getLeastValuablePiece(attackersBitBoard, color, attackerFigType);
-    } while(attackFromBitBoard);
-
-    while (--depth)
-        scores[depth - 1] = - std::max(-scores[depth - 1], scores[depth]);
-    return scores[0];
-}
-
-uint64_t BestMoveSearch::_getLeastValuablePiece(uint64_t pieces, int color, int &pieceIndOut) const {
-    const int range = static_cast<int>(kingIndex);
-    for (int ind = 0; ind < range; ++ind)
-    {
-        const int colIndex = color * static_cast<int>(Board::BitBoardsPerCol) + ind;
-        const uint64_t intersection = pieces & _board.BitBoards[colIndex];
-
-        if (intersection) {
-            pieceIndOut = ind;
-            return ExtractLsbBit(intersection);
-        }
-    }
-
-    return 0;
-}
